requirements.txt
ipykernel
pdfplumber
python-docx
scikit-learn
python-dotenv
langchain
langchain-groq
streamlit